# swagger.json
{
  "swagger": "2.0",
  "info": {
    "description": "Built by team E",
    "version": "1.0.0",
    "title": "Movie App",
    "contact": {
      "email": "romadhonatatang@gmail.com"
    }
  },
  "host": "",
  "basePath": "/api",
...
  "paths": {
...
    "/user/resetPassword/{token}": {
      "put": {
        "tags": [
          "User Collection"
        ],
        "summary": "change password",
        "description": "",
        "operationId": "resetPassword",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "token",
            "in": "path",
            "description": "token of user",
            "required": true,
            "type": "integer",
            "format": "int64"
          },
          {
            "in": "body",
            "name": "body",
            "description": "",
            "required": true,
            "schema": {
              "$ref": "#/definitions/resetPassword"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "password changed",
            "schema": {
              "type": "object",
              "properties": {
                "status": {
                  "type": "boolean",
                  "example": true
                },
                "message": {
                  "type": "string"
                },
                "data": {
                  "type": "object"
                }
              }
            }
          },
          "422": {
            "description": "ERROR",
            "schema": {
              "type": "object",
              "properties": {
                "status": {
                  "type": "boolean",
                  "example": false
                },
                "message": {
                  "type": "string"
                },
                "errors": {
                  "type": "object"
                }
              }
            }
          }
        }
      }
    },
...
    "/user/userProfile": {
      "get": {
        "tags": [
          "User Collection"
        ],
        "summary": "Get user data",
        "description": "",
        "operationId": "userProfile",
        "responses": {
          "200": {
            "description": "success",
            "schema": {
              "type": "object",
              "properties": {
                "status": {
                  "type": "boolean",
                  "example": true
                },
                "message": {
                  "type": "string"
                },
                "data": {
                  "type": "object"
                }
              }
            }
          },
          "422": {
            "description": "ERROR",
            "schema": {
              "type": "object",
              "properties": {
                "status": {
                  "type": "boolean",
                  "example": false
                },
                "message": {
                  "type": "string"
                },
                "errors": {
                  "type": "object"
                }
              }
            }
          }
        },
        "security": [
          {
            "Authorization": []
          }
        ]
      }
    },
...
    "/movie/{movieId}": {
      "get": {
        "tags": [
          "Movie Collection"
        ],
        "summary": "Get movie details",
        "description": "",
        "operationId": "movieDetails",
        "parameters": [
          {
            "name": "movieId",
            "in": "path",
            "description": "movie id",
            "required": true,
            "type": "integer",
            "format": "int64"
          }
        ],
        "responses": {
          "200": {
            "description": "if the email registered, we will sent you email",
            "schema": {
              "type": "object",
              "properties": {
                "status": {
                  "type": "boolean",
                  "example": true
                },
                "message": {
                  "type": "string"
                },
                "data": {
                  "type": "object"
                }
              }
            }
          },
          "422": {
            "description": "ERROR",
            "schema": {
              "type": "object",
              "properties": {
                "status": {
                  "type": "boolean",
                  "example": false
                },
                "message": {
                  "type": "string"
                },
                "errors": {
                  "type": "object"
                }
              }
            }
          }
        },
        "security": [
          {
            "Authorization": []
          }
        ]
      },
      "put": {
        "tags": [
          "Movie Collection"
        ],
        "summary": "Update movie details",
        "description": "",
        "operationId": "updateMovie",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "movieId",
            "in": "path",
            "description": "movie id",
            "required": true,
            "type": "integer",
            "format": "int64"
          },
          {
            "in": "body",
            "name": "body",
            "description": "You can update even if only one parameter",
            "required": true,
            "schema": {
              "$ref": "#/definitions/movies"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "success",
            "schema": {
              "type": "object",
              "properties": {
                "status": {
                  "type": "boolean",
                  "example": true
                },
                "message": {
                  "type": "string"
                },
                "data": {
                  "type": "object"
                }
              }
            }
          },
          "422": {
            "description": "ERROR",
            "schema": {
              "type": "object",
              "properties": {
                "status": {
                  "type": "boolean",
                  "example": false
                },
                "message": {
                  "type": "string"
                },
                "errors": {
                  "type": "object"
                }
              }
            }
          }
        },
        "security": [
          {
            "Authorization": []
          }
        ]
      },
      "delete": {
        "tags": [
          "Movie Collection"
        ],
        "summary": "delete movies",
        "description": "",
        "operationId": "deleteMovie",
        "consumes": [
          "multipart/form-data"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "movieId",
            "in": "path",
            "description": "movie id",
            "required": true,
            "type": "integer",
            "format": "int64"
          }
        ],
        "responses": {
          "200": {
            "description": "if the email registered, we will sent you email",
            "schema": {
              "type": "object",
              "properties": {
                "status": {
                  "type": "boolean",
                  "example": true
                },
                "message": {
                  "type": "string"
                },
                "data": {
                  "type": "object"
                }
              }
            }
          },
          "422": {
            "description": "ERROR",
            "schema": {
              "type": "object",
              "properties": {
                "status": {
                  "type": "boolean",
                  "example": false
                },
                "message": {
                  "type": "string"
                },
                "errors": {
                  "type": "object"
                }
              }
            }
          }
        },
        "security": [
          {
            "Authorization": []
          }
        ]
      }
    }
  },
  "securityDefinitions": {
    "Authorization": {
      "type": "apiKey",
      "name": "Authorization",
      "in": "header"
    }
  },
  "definitions": {
...
    "resetPassword": {
      "type": "object",
      "required": [
        "email"
      ],
      "properties": {
        "password": {
          "type": "string",
          "example": "aandeani"
        },
        "password_confirmation": {
          "type": "string",
          "example": "aandeani"
        }
      }
    },
...
    "movies": {
      "type": "object",
      "required": [
        "title",
        "synopsis"
      ],
      "properties": {
        "title": {
          "type": "string",
          "example": "A Star Is Born"
        },
        "releaseYears": {
          "type": "number",
          "example": 2018
        },
        "genre": {
          "type": "string",
          "example": "Drama"
        },
        "duration": {
          "type": "number",
          "example": 134
        },
        "trailer": {
          "type": "string",
          "example": "https://www.youtube.com/watch?v=nSbzyEJ8X9E"
        },
        "synopsis": {
          "type": "string",
          "example": "A musician helps a young singer find fame as age and alcoholism send his own career into a downward spiral."
        }
      }
    }
  },
...
}